Split the carousel into multiple subcomponents
